mankatha database modified

+++ apps/mankatha/delw.php
@@ -1,9 +1,24 @@
 <?php
 		session_start();
-include('../../connect.php');
+$connection=mysql_connect("10.0.0.126","festember11","festember");
+
+if(!$connection){
+
+  die("Database connection failed:" . mysql_error());
+
+}
+
+$db_select=mysql_select_db("delta",$connection);
+
+if(!$db_select){
+
+  die("Database connection failed:" . mysql_error());
+
+}
+
 	$query="DELETE FROM mankatha_random";
-	mysql_query($query);
-    mysql_close();
+mysql_query($query,$connection);
+    mysql_close($connection);
 ?>
 <html>
